test : test코드 작성완료

+++ src/calculator/MyCalculator.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class MyCalculator {
     public double root(String command){
-        int x = Integer.parseInt(command);
+        double x = Double.parseDouble(command);
         return (double) Math.round(Math.sqrt(x) * 100) / 100;
     }
